refactor: rename upe-styles/ directory to styles/
The client/checkout/upe-styles/ directory serves both UPE and WooPay
style extraction. The previous name implied UPE-only scope, which no
longer reflects actual usage.

- Rename client/checkout/upe-styles/ to client/checkout/styles/
- Update all 12 import paths across source and test files
- No behavior change — pure refactor

Fixes #11448
