+++ client/checkout/styles/__tests__/woopay-appearance.test.js
